component needs to be in addon/components to be extendable in the usual way. I've also needed to upgrade ember in order to get things to build.
