+++ ember-cli-build.js
@@ -1,0 +1,8 @@
+var EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
+
+module.exports = function(defaults) {
+  var app = new EmberAddon(defaults, {
+    // Any other options
+  });
+  return app.toTree();
+};
